core: switch to hooking vmachinecommands, part 3/?
